Remove node runtime and replace with deno
Done to reduce maintenance cost, and because of deno's bundling to js
capabilities to support browser usecases.
