+++ serde-generate/runtime/typescript/bincode/bincodeDeserializer.ts
@@ -1,16 +1,18 @@
-import {BinaryDeserializer} from "../serde/binaryDeserializer";
+import { BinaryDeserializer } from "../serde/binaryDeserializer.ts";
 
 export class BincodeDeserializer extends BinaryDeserializer {
-    deserializeLen(): number {
-        return Number(this.deserializeU64());
-    }
+  deserializeLen(): number {
+    return Number(this.deserializeU64());
+  }
 
-    public deserializeVariantIndex(): number {
-        return this.deserializeU32();
-    }
+  public deserializeVariantIndex(): number {
+    return this.deserializeU32();
+  }
 
-    checkThatKeySlicesAreIncreasing(key1: [number, number], key2: [number, number]): void {
-        return;
-    }
-
+  checkThatKeySlicesAreIncreasing(
+    key1: [number, number],
+    key2: [number, number],
+  ): void {
+    return;
+  }
 }
