+++ serde-generate/runtime/typescript/bincode/bincodeSerializer.ts
@@ -1,15 +1,15 @@
-import {BinarySerializer} from "../serde/binarySerializer";
+import { BinarySerializer } from "../serde/binarySerializer.ts";
 
 export class BincodeSerializer extends BinarySerializer {
-    serializeLen(value: number): void {
-        this.serializeU64(value);
-    }
+  serializeLen(value: number): void {
+    this.serializeU64(value);
+  }
 
-    public serializeVariantIndex(value: number): void {
-        this.serializeU32(value);
-    }
+  public serializeVariantIndex(value: number): void {
+    this.serializeU32(value);
+  }
 
-    public sortMapEntries(offsets: number[]): void {
-        return;
-    }
+  public sortMapEntries(offsets: number[]): void {
+    return;
+  }
 }
