--- serde-generate/runtime/typescript/bincode/index.ts
@@ -1,2 +1,0 @@
-export {BincodeSerializer} from "./bincodeSerializer";
-export {BincodeDeserializer} from "./bincodeDeserializer";
