+++ serde-generate/runtime/typescript/serde/types.ts
@@ -1,7 +1,7 @@
 export type Optional<T> = T | null;
 export type Seq<T> = T[];
-export type Tuple<T extends any[]> = T
-export type ListTuple<T extends any[]> = Tuple<T>[]
+export type Tuple<T extends any[]> = T;
+export type ListTuple<T extends any[]> = Tuple<T>[];
 
 export type unit = null;
 export type bool = boolean;
