+++ serde-generate/tests/typescript_generation.rs
@@ -1,127 +1,91 @@
 // Copyright (c) Facebook, Inc. and its affiliates
 // SPDX-License-Identifier: MIT OR Apache-2.0
 
+use regex::Regex;
 use serde_generate::{test_utils, typescript, CodeGeneratorConfig, Encoding, SourceInstaller};
-use std::{
-    fs::File,
-    io::{Result, Write},
-    process::Command,
-};
-use tempfile::{tempdir, TempDir};
+use std::{fs::File, path::Path, process::Command};
+use tempfile::tempdir;
 
-fn write_package_tsconfig_json_for_test_build(path: std::path::PathBuf) -> Result<()> {
-    let mut package_json = std::fs::File::create(path.join("package.json"))?;
-    writeln!(
-        package_json,
-        r#"
-{{
-  "name": "tmpCode",
-  "version": "1.0.0",
-  "main": "dist/index.js",
-  "types": "dist/index.d.ts",
-  "scripts": {{
-    "build": "tsc"
-  }},
-  "dependencies": {{
-  }},
-  "devDependencies": {{
-    "@types/node": "12.12.2",
-    "typescript": "^3.9.6"
-  }}
-}}
-"#
-    )?;
-    let mut tsconfig_json = std::fs::File::create(path.join("tsconfig.json"))?;
-    writeln!(
-        tsconfig_json,
-        r#"
-{{
-  "compilerOptions": {{
-    "target": "es6",
-    "module": "commonjs",
-    "declaration": true,
-    "outDir": "./dist",
-    "strict": true,
-    "esModuleInterop": true,
-    "skipLibCheck": true,
-    "forceConsistentCasingInFileNames": true,
-    "lib": [
-      "es6",
-      "esnext.BigInt",
-      "dom"
-    ]
-
-  }},
-  "include": ["testing/*.ts"]
-}}
-"#
-    )?;
-    Ok(())
-}
-
-fn test_that_ts_code_compiles_with_config(
+fn test_typescript_code_compiles_with_config(
+    dir_path: &Path,
     config: &CodeGeneratorConfig,
-) -> (TempDir, std::path::PathBuf) {
+) -> std::path::PathBuf {
     let registry = test_utils::get_registry().unwrap();
-    let dir = tempdir().unwrap();
-    make_output_file(&dir);
-    let source_path = dir.path().join("testing").join("test.ts");
+    make_output_file(dir_path);
+
+    let installer = typescript::Installer::new(dir_path.to_path_buf());
+    installer.install_serde_runtime().unwrap();
+    assert_deno_info(dir_path.join("bcs/mod.ts").as_path());
+
+    installer.install_bcs_runtime().unwrap();
+    assert_deno_info(dir_path.join("serde/mod.ts").as_path());
+
+    let source_path = dir_path.join("testing").join("test.ts");
     let mut source = File::create(&source_path).unwrap();
 
     let generator = typescript::CodeGenerator::new(config);
     generator.output(&mut source, &registry).unwrap();
-    let _result = write_package_tsconfig_json_for_test_build(dir.path().to_path_buf());
 
-    let installer = typescript::Installer::new(dir.path().to_path_buf());
-    installer.install_serde_runtime().unwrap();
-    installer.install_bincode_runtime().unwrap();
-    installer.install_bcs_runtime().unwrap();
-
-    let npm_status = Command::new("npm")
-        .arg("install")
-        .current_dir(dir.path())
-        .status()
-        .unwrap();
-    assert!(npm_status.success());
-
-    let status = Command::new("npm")
-        .arg("run")
-        .arg("build")
-        .current_dir(dir.path())
-        .status()
-        .unwrap();
-    assert!(status.success());
-
-    let path = dir.path().join("testing");
-    (dir, path)
+    assert_deno_info(&source_path);
+    dir_path.join("testing")
 }
 
-fn make_output_file(dir: &TempDir) {
-    std::fs::create_dir_all(dir.path().join("testing")).unwrap_or(());
+fn assert_deno_info(ts_path: &Path) {
+    let output = Command::new("deno")
+        .arg("info")
+        .arg(ts_path)
+        .output()
+        .expect("deno info failed, is deno installed? brew install deno");
+    assert!(output.status.success());
+
+    let stdout = String::from_utf8(output.stdout).unwrap();
+    assert!(
+        !is_error_output(stdout.as_str()),
+        "deno info detected an error\n{}",
+        stdout
+    );
 }
 
-#[test]
-fn test_that_ts_code_compiles() {
-    let config = CodeGeneratorConfig::new("testing".to_string());
-    test_that_ts_code_compiles_with_config(&config);
+fn is_error_output(output: &str) -> bool {
+    let re = Regex::new(r"\berror\b").unwrap();
+    re.is_match(output)
+}
+
+fn make_output_file(dir: &Path) {
+    std::fs::create_dir_all(dir.join("testing")).unwrap_or(());
 }
 
 #[test]
-fn test_that_ts_code_compiles_with_bcs() {
+fn test_is_error_output() {
+    let table = vec![
+        ("file:///var/folders/l0/x592_pjj18n6r2m0nqn05vmc0000gn/T/.tmp5NPlE2/serde/mod.ts (176B)", false),
+        ("https://deno.land/std@0.85.0/node/_errors.ts (60.89KB)", false),
+        ("error: Cannot resolve module \"file:///var/folders/l0/x592_pjj18n6r2m0nqn05vmc0000gn/T/.tmp5NPlE2/bcs/mod.ts\"", true),
+        ("file:///var/folders/l0/x592_pjj18n6r2m0nqn05vmc0000gn/T/.tmpG1an6c/something/noSerializer.ts (error)", true),
+    ];
+
+    for (input, expectation) in table {
+        assert_eq!(is_error_output(input), expectation);
+    }
+}
+
+#[test]
+fn test_typescript_code_compiles_with_bcs() {
+    let dir = tempdir().unwrap();
     let config =
         CodeGeneratorConfig::new("testing".to_string()).with_encodings(vec![Encoding::Bcs]);
-    test_that_ts_code_compiles_with_config(&config);
+    test_typescript_code_compiles_with_config(&dir.path(), &config);
 }
 
 #[test]
-fn test_that_ts_code_compiles_with_comments() {
+fn test_typescript_code_compiles_with_comments() {
+    let dir = tempdir().unwrap();
     let comments = vec![(vec!["SerdeData".to_string()], "Some\ncomments".to_string())]
         .into_iter()
         .collect();
     let config = CodeGeneratorConfig::new("testing".to_string()).with_comments(comments);
 
-    let (_dir, path) = test_that_ts_code_compiles_with_config(&config);
-
+    let path = test_typescript_code_compiles_with_config(&dir.path(), &config);
     // Comment was correctly generated.
     let content = std::fs::read_to_string(path.join("test.ts")).unwrap();
     assert!(content.contains(
